-corrigido modal de edicao (lojista e produto) e exclusao (produto) -Formatacao css lojistaLogado.php -imagens estao sendo exluidas do arquivo ao editar ou excluir produtos

+++ restrito/excluirScript.php
@@ -1,19 +1,36 @@
 <?php
 include "../validarLojista.php";
 include "conexao.php";
+include "../scripts.php";
+
 $conn = pegarConexao('lojista');
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    $id = $_POST['id'];
-    $nome = $_POST['nome'];
+    $id = clear($conn, $_POST['id']);
     
-    $SQL = "DELETE FROM produtos WHERE id = '$id'"; 
+    // Obter o caminho da imagem do produto antes de deletar
+    $queryImagem = "SELECT imagem, imagem2, imagem3 FROM produtos WHERE id = '$id'";
+    $result = $conn->query($queryImagem);
 
-    if ($conn->query($SQL) === TRUE) {
-        unlink($imagem);
-        header('Location: lojistaLojista.php');
+    if ($result->num_rows > 0) {
+        $row = $result->fetch_assoc();
+        $imagem = '../img/' . $row['imagem'];
+        $imagem2 = '../img/' . $row['imagem2'];
+        $imagem3 = '../img/' . $row['imagem3'];
+
+        // Excluir o produto
+        $SQL = "DELETE FROM produtos WHERE id = '$id'"; 
+
+        if ($conn->query($SQL) === TRUE) {
+            unlink($imagem);
+            unlink($imagem2);
+            unlink($imagem3);
+            header('Location: lojistaLojista.php');
+        } else {
+            echo "Erro ao excluir item: " . $conn->error;
+        }
     } else {
-        echo "Erro ao excluir item: " . $conn->error;
+        echo "Produto não encontrado.";
     }
 } else {
     echo "Requisição inválida para exclusão";
